refactor(ui): seperate file for each store
