+++ ui/src/AttachmentsDialog.svelte
@@ -3,7 +3,7 @@
   import { cloneDeep } from "lodash";
   import type { Board, Card } from "./board";
   import { getContext } from "svelte";
-  import type { KanDoStore } from "./store";
+  import type { KanDoStore } from "./stores/kando";
   import { type WALUrl} from "./util";
   import '@shoelace-style/shoelace/dist/components/button/button.js';
   import '@shoelace-style/shoelace/dist/components/dialog/dialog.js';
