+++ ui/src/BoardMenu.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { getContext } from "svelte";
-    import { KanDoCloneManagerStore, USING_FEEDBACK, type KanDoStore } from "./store";
+    import { KanDoCloneManagerStore, USING_FEEDBACK, type KanDoStore } from "./stores/kando";
     import { type EntryHash } from '@holochain/client';
     import GroupParticipants from './GroupParticipants.svelte';
     import NewBoardDialog from './NewBoardDialog.svelte';
