+++ ui/src/FeedbackItems.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { createEventDispatcher, getContext } from "svelte";
-  import type { KanDoStore } from "./store";
+  import type { KanDoStore } from "./stores/kando";
   import type { EntryHash } from "@holochain/client";
   import "@shoelace-style/shoelace/dist/components/skeleton/skeleton.js";
   import type { BoardState } from "./board";
