+++ ui/src/FeedbackPane.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { getContext, onMount } from "svelte";
-  import type { KanDoStore } from "./store";
+  import type { KanDoStore } from "./stores/kando";
   import type { BoardAndLatestState } from "./boardList";
   import {
     get,
