+++ ui/src/KanDoPane.svelte
@@ -3,7 +3,7 @@
   import CardEditor from "./CardEditor.svelte";
   import CardDetailsDrawer from "./CardDetailsDrawer.svelte";
   import EmojiIcon from "./EmojiIcon.svelte";
-  import type { KanDoStore } from "./store";
+  import type { KanDoStore } from "./stores/kando";
   import LabelSelector from "./LabelSelector.svelte";
   import { v1 as uuidv1 } from "uuid";
   import { type Card, Group, UngroupedId, type CardProps, type Comment, type Checklists, Board, type BoardProps, feedItemsGroupedByCard, MAX_FEED_ITEMS, UngroupedName } from "./board";
