+++ ui/src/stores/kando.ts
@@ -2,31 +2,23 @@ import {
     type AppClient,
     type EntryHash,
     type AgentPubKeyB64,
-    type AppCallZomeRequest,
     type RoleName,
     encodeHashToBase64,
-    type EntryHashB64,
     type AgentPubKey,
     decodeHashFromBase64,
     type Timestamp,
     type DnaHash,
-    type CellInfo,
-    CellType,
-    type CellId,
 } from '@holochain/client';
 import { SynStore,  SynClient} from '@holochain-syn/core';
-import { BoardList } from './boardList';
+import { BoardList } from '../boardList';
 import TimeAgo from "javascript-time-ago"
 import en from 'javascript-time-ago/locale/en'
 import type { v1 as uuidv1 } from "uuid";
 import { derived, get, writable, type Unsubscriber, type Writable } from "svelte/store";
-import { ProfilesClient, ProfilesStore } from '@holochain-open-dev/profiles';
-import { UngroupedName, type BoardState } from './board';
+import { ProfilesStore } from '@holochain-open-dev/profiles';
+import { UngroupedName, type BoardState } from '../board';
 import type { WeaveClient } from '@lightningrodlabs/we-applet';
 import { HoloHashMap } from '@holochain-open-dev/utils';
-import { v7 as uuidv7 } from "uuid";
-import { asyncDerived, type Loadable } from '@square/svelte-store';
-import { hashEqual } from './util';
 
 // @ts-ignore
 export const USING_FEEDBACK :boolean | undefined = window.__USING_FEEDBACK || (import.meta as any).env.VITE_USING_FEEDBACK
@@ -247,148 +239,4 @@ export class KanDoStore {
         return "Unknown"
     }
 
-}
-
-export interface CellInfoNormalized {
-    cellId: CellId;
-    cellInfo: CellInfo;
-    roleName: string;
-    name: string;
-    networkSeed: string;
-    displayName: string;
-}
-
-export class KanDoCloneManagerStore {
-    activeDnaHash: Writable<DnaHash>;
-    activeCellInfoNormalized: Loadable<CellInfoNormalized>;
-    activeStore: Loadable<KanDoStore>;
-
-    constructor(
-        public client: AppClient,
-        public weaveClient?: WeaveClient,
-    ) {
-        this.activeDnaHash = writable<DnaHash>();
-        this.activeDnaHash.subscribe(this._saveActiveDnaHash);
-        this.activeCellInfoNormalized = asyncDerived(this.activeDnaHash, async ($activeDnaHash) => {
-            const appInfo = await this.client.appInfo();
-
-            if(!$activeDnaHash) {
-                await this._loadActiveDnaHash();
-                $activeDnaHash = get(this.activeDnaHash);
-            }
-            
-            const cellInfo = appInfo.cell_info[ROLE_NAME].find((cellInfo: CellInfo) => {
-                if(CellType.Provisioned in cellInfo) {
-                    return hashEqual(cellInfo[CellType.Provisioned].cell_id[0], $activeDnaHash);
-                } else if(CellType.Cloned in cellInfo) {
-                    return hashEqual(cellInfo[CellType.Cloned].cell_id[0], $activeDnaHash);
-                }
-            });
-
-            if(CellType.Provisioned in cellInfo) {
-                return {
-                    cellId: cellInfo[CellType.Provisioned].cell_id,
-                    cellInfo: cellInfo,
-                    roleName: ROLE_NAME,
-                    name: cellInfo[CellType.Provisioned].name,
-                    networkSeed: cellInfo[CellType.Provisioned].dna_modifiers.network_seed,
-                    displayName: cellInfo[CellType.Provisioned].dna_modifiers.network_seed === "" ? "Public" : cellInfo[CellType.Provisioned].name,
-                };
-            } else if(CellType.Cloned in cellInfo) {
-                return {
-                    cellId: cellInfo[CellType.Cloned].cell_id,
-                    cellInfo: cellInfo,
-                    roleName: cellInfo[CellType.Cloned].clone_id,
-                    name: cellInfo[CellType.Cloned].name,
-                    networkSeed: cellInfo[CellType.Cloned].dna_modifiers.network_seed,
-                    displayName: cellInfo[CellType.Cloned].dna_modifiers.network_seed === "" ? "Public" : cellInfo[CellType.Cloned].name,
-                };
-            }
-        });
-        this.activeStore = asyncDerived([this.activeDnaHash, this.activeCellInfoNormalized], async ([$activeDnaHash, $activeCellInfoNormalized]) => {
-            await this.activeCellInfoNormalized.load();
-
-            const profilesClient = this.weaveClient !== undefined ? weaveClient.renderInfo.profilesClient : new ProfilesClient(this.client, $activeCellInfoNormalized.roleName);
-            return new KanDoStore(this, new ProfilesStore(profilesClient), $activeDnaHash, $activeCellInfoNormalized.roleName);
-        });
-        this._loadActiveDnaHash();
-    }
-    
-    async list(): Promise<CellInfoNormalized[]> {
-        const appInfo = await this.client.appInfo();
-        const cells = appInfo.cell_info[ROLE_NAME];
-        let cellsNormalized =  cells.map((cell) => {
-            if(CellType.Provisioned in cell) {
-                return {
-                    cellId: cell[CellType.Provisioned].cell_id, 
-                    cellInfo: cell,
-                    roleName: ROLE_NAME,
-                    name: cell[CellType.Provisioned].name,
-                    networkSeed: cell[CellType.Provisioned].dna_modifiers.network_seed,
-                    displayName: cell[CellType.Provisioned].dna_modifiers.network_seed === "" ? "Public" : cell[CellType.Provisioned].name,
-                };
-            } else if(CellType.Cloned in cell) {
-                return {
-                    cellId: cell[CellType.Cloned].cell_id,
-                    cellInfo: cell,
-                    roleName: cell[CellType.Cloned].clone_id,
-                    name: cell[CellType.Cloned].name,
-                    networkSeed: cell[CellType.Cloned].dna_modifiers.network_seed,
-                    displayName: cell[CellType.Cloned].dna_modifiers.network_seed === "" ? "Public" : cell[CellType.Cloned].name,
-                };
-            }
-        });
-        cellsNormalized.sort((a,b) => a.networkSeed < b.networkSeed ? -1 : 1);
-
-        return cellsNormalized;
-    }
-
-    create(name: string) {
-        return this.client.createCloneCell({
-            name,
-            role_name: ROLE_NAME,
-            modifiers: {
-                network_seed: uuidv7(),
-            }
-        });
-    }
-
-    join(name: string, networkSeed: string) {
-        return this.client.createCloneCell({
-            name,
-            role_name: ROLE_NAME,
-            modifiers: {
-                network_seed: networkSeed 
-            }
-        });
-    }
-
-    disable(cellId: CellId) {
-        return this.client.disableCloneCell({ clone_cell_id: cellId });
-    }
-
-    enable(cellId: CellId) {
-        return this.client.enableCloneCell({ clone_cell_id: cellId })
-    }
-    
-    activate(cellId: CellId) {
-        this.activeDnaHash.set(cellId[0]);
-    }
-
-    private async _loadActiveDnaHash() {
-        const dnaHash = localStorage.getItem("activeDnaHash");
-        if(dnaHash !== null && dnaHash !== undefined) {
-            this.activeDnaHash.set(decodeHashFromBase64(dnaHash));
-        } else {
-            const appInfo = await this.client.appInfo();
-            const defaultDnaHash = appInfo.cell_info[ROLE_NAME][0][CellType.Provisioned].cell_id[0];
-            this.activeDnaHash.set(defaultDnaHash);
-        }
-    }
-
-    private _saveActiveDnaHash(val: DnaHash) {
-        if(val !== undefined && val !== null) {
-            localStorage.setItem("activeDnaHash", encodeHashToBase64(val));
-        }
-    }
 }
